+++ R/boxGrobs_connect_pr_one_to_many_boxes.R
@@ -1,4 +1,4 @@
-prConnectMany <- function(
+prConnectOneToMany <- function(
     start,
     ends,
     type,